Renamed WorldObject -> WorldEntity in Catch.
