+++ Catch/src/WorldEntities/scorearea.cpp
